+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [#84](https://github.com/green-code-initiative/creedengo-javascript/pull/84) Add rule GCI535 "No imported number format library"
+- [#118](https://github.com/green-code-initiative/creedengo-javascript/pull/118) Add rule GCI2536 "Move the browserslist configuration to a production target in package.json"
 
 ## [3.1.0] - 2026-05-10